Mention tip for users with lack of .xcscheme file
After banging my head against this for hours last night, I thought it might be helpful to leave this tip for any other unlucky folks.  Probably a rare situation (mine was caused by a merge), but I was doing what the docs suggested, and Carthage was just not building my frameworks.